change id to integer

+++ src/main/java/com/example/demo/repositories/CourseRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CourseRepository extends JpaRepository<CourseEntity, Long> {
+public interface CourseRepository extends JpaRepository<CourseEntity, Integer> {
     Optional<CourseEntity> findByName(String name);
+
 }
